requirements.txt
Django==1.8
django-autofixture==0.10.0
django-dotenv==1.3.0
django-filter==0.9.2
django-getenv==1.3.1
django-gunicorn==0.1.1
djangorestframework==3.1.1
gunicorn==19.3.0
Markdown==2.6.2
Pillow==2.8.1
psycopg2==2.6
